Try to detect and ignore multisegmented or resumed downloads
Download managers such as IDM often connect to a server multiple times in order to more quickly download a file in segments.  Tools such as wget can also resume an interupted download later.  If possible, we don't want to count any of this behaviour, and only count unique downloads once.

The implementation as it stands isn't perfect - it currently ignores requests which have either a Range or Referer header, which ends up excluding all downloads started in a download manager.  However, downloads which originate in the browser are counted.