Implement a basic token-based guard against counting duplicate downloads
Download tokens are created using a base64 encoding of a file's numeric ID, the timestamp of when they were issued and a user's IP address.  This is then stored in a file-specific cookie on the user's machine.  The /download endpoint decodes these tokens and ensures that they are still valid, i.e. that the file ID and IP address match, and that the token wasn't issued too long ago.

The period for which tokens are valid is configurable using the TOKEN_VALIDITY_PERIOD setting and is 600 seconds by default.  An HTTP/403 is returned if a /download endpoint is accessed without an appropriate token.